Renamed the webhook-event endpoint class and moved to relevant package
