+++ webhook-service/conf/routes
@@ -1,4 +1,4 @@
 GET     /ping                 ch.datascience.webhookservice.HealthCheckController.ping
 
-POST    /webhook-event        ch.datascience.webhookservice.PushEventConsumer.processPushEvent
+POST    /webhook-event        ch.datascience.webhookservice.eventprocessing.WebhookEventEndpoint.processPushEvent
 POST    /projects/:id/hooks   ch.datascience.webhookservice.hookcreation.HookCreationEndpoint.createHook(id: ProjectId)
